move out the content of tres_marias folder
